TInkering with dialog system
I want to get all the dialog text into one file for easy reference and possible
localizaiton concerns, and ideally with nice editor support (i.e. a dropdown to
select which key to use instead of having to type it in).  This is actually mostly
working, but as it stands I need a DialogEntry to reference the key, which is
kindof double work - havve some notes on how I might work around that.