Finished Dialog Tree System
Dialog Entries are now automatically populated by the DialogTree when you add
keys to the DialogTree's lookup table; each level will subclass DialogTree to
set a customer lookup table for text keys to dialog contents.  Dialog Entry nodes
are now named for their key in the lookup table, and are provide a few features:

* Easy, in-editor relationship with TextBoxTriggers/the trigger system through
  node paths in the scene tree
* Metadata on top of text (i.e. dialog style, profile pictures, etc) managed
  through the scnee tree
* Easy assocation to/lookup of actual text box content

The automatic creation  of children in the scene tree in editor allows for
the best of both worlds - all text can be added in one single file that is easy
to search and audit, and the scene tree can have an entry per available dialog
text, but without having to manually associate the two.