Added the concept of "Room Layers"
To simplify the setup of the tutorial and other possible special game states,
a "Room Layer" can take the graphics and itneractions of a normal LevelRoom, but
extend them with its own.  This is useful to be able to reuse common components
between the tutorial segment and the main game while not requiring two versions
of the same room to exist.