Started on item interaction
The GameController will track the active item and emit signals to let InventoryItems
know if they should be highlighted or not.  Actual highlighting coming soon.

Additionally, Clickables that interact with items wil have an "item" trigger; this
will eventualy be set to an "ItemBehaviorTrigger" which can pass execution to a
Trigger chain in the Item, and an `item_fire` method now exists to allow items to
pass execution back to the Clickable where this makes sense (i.e. is more than just
a text box).