Item interactions, better living room, etc
Fixed the ladder animation (going up), item selection now has a display
element, and more.

Items interact with Clickables using the "item" trigger, which currently acts as
any other trigger in the chain.  In the future there'll be a specific item trigger
that uses the item's name key to decide how to behave.

Item dismissal is still a challenge - _unhandled_input is called too early to dismiss
the active item _after_ Clickable's _on_Area2D_input_event is called, so I'm using
an "ItemResetFilter," which is just a large Area2D that sits behind everything and
clears the active item when clicked.  This works, but sometimes it's called before
a Clickable is.. the docs suggest that keeping it at the top of the root scene
tree should ensure it's always last, but in practice it isn't.  Still looking at
why.